requirement.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-extensions==2.0.7
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
mysqlclient==1.3.12
oauthlib==2.1.0
PyJWT==1.6.4
PyMySQL==0.8.1
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
